# Conflicts:
#	salience_sum_crf/reader/summ_data_reader.py

+++ salience_sum_crf/reader/summ_data_reader.py
@@ -40,7 +40,7 @@ class SummDataReader(DatasetReader):
             cs = CubicSpline(list(value_dict.keys()), list(value_dict.values()))
         except:
             return value
-        c = [float(cs(i)) if i in value_dict.keys() else cs(i) * cs(i) for i in range(len(value))]
+        c = [int(cs(i)) if i in value_dict.keys() else cs(i) * cs(i) for i in range(len(value))]
         return c
 
     def text_to_instance(self, src_seq: List[Token], tgt_seq: str, salience_seq: List[str]) -> Instance:
